+++ lib/redarx.dart
@@ -4,6 +4,9 @@
 /// Redarx : action dispatcher commander store
 library redarx;
 
+export 'src/commander-config.dart';
+export 'src/commander.dart';
+export 'src/commands-reducer-transformer.dart';
 export 'src/commands.dart';
 export 'src/dispatcher.dart';
 export 'src/model.dart';