+++ lib/src/dispatcher.dart
@@ -4,11 +4,13 @@ import 'package:redarx/src/request.dart';
 
 /// actions dispatcher
 class Dispatcher {
-  StreamController<Request> streamr = new StreamController();
+
+  /// stream of command requests : requests are added via the dispatch() method
+  final StreamController<Request> streamController = new StreamController();
 
   /// dispatched [Stream]<[Request]>
-  Stream<Request> get onRequest => streamr.stream;
+  Stream<Request> get request$ => streamController.stream;
 
-  /// add [Request]
-  dispatch(Request a) => streamr.add(a);
+  /// add [Request] to stream
+  void dispatch(Request a) => streamController.add(a);
 }