+++ lib/src/model.dart
@@ -4,6 +4,8 @@ typedef S InitialStateProvider<S extends AbstractModel>();
 /// state Model base class
 abstract class AbstractModel {
 
+  const AbstractModel();
+
   /// provide an instance of the model in initialState
   AbstractModel initial();
 }