+++ lib/src/store.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:meta/meta.dart';
+import 'package:redarx/src/commands-reducer-transformer.dart';
 import 'package:redarx/src/commands.dart';
 import 'package:redarx/src/model.dart';
 
@@ -35,71 +36,5 @@ class Store<C extends Command<M>, M extends AbstractModel> {
     state$ = historyStream.transform(
         new CommandStreamReducer<Command<M>, M>.broadcast(
             initialStateProvider().initial() as M));
-  }
-}
-
-/**
- * Stream transformer : reduce state from commands stream
- * transform a Stream<Command> en Stream<AbstractModel>
- */
-class CommandStreamReducer<S extends Command, T extends AbstractModel>
-    implements StreamTransformer<S, T> {
-  T state;
-
-  List<S> history;
-
-  StreamController<T> _controller;
-
-  bool cancelOnError;
-
-  StreamSubscription _subscription;
-
-  Stream<S> _stream;
-
-  CommandStreamReducer(this.state, {bool sync: false, this.cancelOnError}) {
-    _controller = new StreamController<T>(
-        onListen: _onListen,
-        onCancel: _onCancel,
-        onPause: () {
-          _subscription.pause();
-        },
-        onResume: () {
-          _subscription.resume();
-        },
-        sync: sync);
-  }
-
-  CommandStreamReducer.broadcast(this.state,
-      {bool sync: false, this.cancelOnError}) {
-    _controller = new StreamController<T>.broadcast(
-        onListen: _onListen, onCancel: _onCancel, sync: sync);
-  }
-
-  void _onListen() {
-    _subscription = _stream.listen(onData,
-        onError: _controller.addError,
-        onDone: _controller.close,
-        cancelOnError: cancelOnError);
-  }
-
-  void _onCancel() {
-    _subscription.cancel();
-    _subscription = null;
-  }
-
-  // on stream data : exec (sync or async) the command
-  // & inject newstate in the state$ stream
-  Future onData(S cmd) async {
-    if (cmd is AsyncCommand)
-      state = await (cmd as AsyncCommand<T>).execAsync(state);
-    else
-      state = (cmd as Command<T>).exec(state);
-    _controller.add(state);
-  }
-
-  @override
-  Stream<T> bind(Stream<S> stream) {
-    this._stream = stream;
-    return _controller.stream;
   }
 }